Represent ByteVectors using Vector[Short]
Vector[Int] is 32bit which is excessive. However, Vector[Byte] is signed
which makes it tricky to deal with as bytevectors are unsigned. Settle
on Vector[Short] which can represent all the values in a bytevector
without sign tricks

+++ compiler/src/main/scala/llambda/AbstractSyntaxTree.scala
@@ -128,7 +128,7 @@ case class VectorLiteral(elements : Vector[Datum]) extends Datum {
     "#(" + elements.map(_.toString).mkString(" ") + ")"
 }
 
-case class Bytevector(elements : Vector[Int]) extends NonSymbolLeaf {
+case class Bytevector(elements : Vector[Short]) extends NonSymbolLeaf {
   override def toString = 
     "#u8(" + elements.map(_.toString).mkString(" ") + ")"
 }
